requirements.txt
customtkinter==5.2.2
opencv-python==4.9.0.80
pillow==10.2.0
aiortc==1.7.0